Add LusasAdapterV191, change all GetId() requests from Lusas to long for compatibility with previous versions

+++ Lusas_Adapter/Convert/ToBHoM/Loads/ToAreaUniformlyDistributedLoad.cs
@@ -77,7 +77,7 @@ namespace BH.Adapter.Adapters.Lusas
                 GetName(lusasDistributed));
             }
 
-            int adapterNameId = lusasDistributed.getID();
+            long adapterNameId = lusasDistributed.getID();
             areaUniformlyDistributedLoad.SetAdapterId(typeof(LusasId), adapterNameId);
 
             return areaUniformlyDistributedLoad;
